Finalize test for circular colorbar
